+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.9.0+dev
 - made constructor of static classes private
+- remove nette/utils from dependencies
 
 Version 1.9.0
 - raised minimal version of PHP to 8.1